@@ -4,6 +4,7 @@
   "ui5": {
     "webComponentsPackage": true
   },
+  "type": "module",
   "scripts": {
     "clean": "wc-dev clean",
     "lint": "wc-dev lint",
@@ -22,14 +23,14 @@
     "./*": "./dist/*"
   },
   "dependencies": {
-    "@ui5/webcomponents": "^1.14.3",
-    "@ui5/webcomponents-base": "1.14.2",
-    "@ui5/webcomponents-icons": "^1.14.3",
-    "@ui5/webcomponents-theming": "1.14.2"
+    "@ui5/webcomponents": "2.8.0",
+    "@ui5/webcomponents-base": "2.8.0",
+    "@ui5/webcomponents-icons": "2.8.0",
+    "@ui5/webcomponents-theming": "2.8.0"
   },
   "devDependencies": {
-    "@ui5/webcomponents-tools": "1.14.2",
+    "@ui5/webcomponents-tools": "2.8.0",
     "chromedriver": "*",
-    "typescript": "^4.9.4"
+    "typescript": "^5.6.2"
   }
 }
